refactor(WinningLottoCalculator) : getRateOfReturn 메서드 리팩토링
정수형 totalPrizeMoney를 실수형으로 리팩토링

+++ src/main/java/domain/WinningLottoCalculator.java
@@ -61,7 +61,7 @@ public class WinningLottoCalculator {
 	}
 
 	public double getRateOfReturn(PurchaseAmount purchaseAmount) {
-		return roundToTwoDecimalPlaces((this.totalPrizeMoney / purchaseAmount.getAmount()) * 100);
+		return roundToTwoDecimalPlaces((((double) this.totalPrizeMoney) / purchaseAmount.getAmount()) * 100);
 	}
 
 	public double roundToTwoDecimalPlaces(double rateOfReturn) {
